ml-service/requirement.txt
# ============================================
# Core Flask Framework
# ============================================
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# ============================================
# Resume Parsing (PDF/DOCX)
# ============================================
pymupdf==1.23.8
python-docx==1.1.0
pdfplumber==0.10.3

# ============================================
# OCR (Optical Character Recognition)
# ============================================
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# ============================================
# NLP & Embeddings
# ============================================
sentence-transformers==2.5.0
transformers==4.36.0
torch==2.1.0
spacy==3.7.2

# ============================================
# Machine Learning & Data Processing
# ============================================
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
lightgbm==4.1.0

# ============================================
# Vector Database Clients
# ============================================
qdrant-client==1.7.0
pinecone-client==3.0.0

# ============================================
# Text Processing & Utilities
# ============================================
nltk==3.8.1
regex==2023.12.25
unidecode==1.3.7
python-multipart==0.0.6

# ============================================
# HTTP & API Utilities
# ============================================
requests==2.31.0
urllib3==2.1.0

# ============================================
# Validation & Schema
# ============================================
pydantic==2.5.3
marshmallow==3.20.1

# ============================================
# Date/Time Processing
# ============================================
python-dateutil==2.8.2
